+++ src/sensors/SensorManager.hpp
@@ -4,7 +4,6 @@
 #include "Sensor.hpp"
 #include "comms/config_layer.hpp"
 #include "d200.hpp"
-#include "dr16.hpp"
 #include "ICM20649.hpp"
 #include <SPI.h>
 #include "buff_encoder.hpp"
