+++ src/utils/wrapping.hpp
@@ -3,10 +3,6 @@
 #ifndef WRAPPING_H
 #define WRAPPING_H
 
-float wrapAngle(float angle) {
-	while (angle >= M_PI) angle -= 2 * M_PI;
-	while (angle <= -M_PI) angle += 2 * M_PI;
-	return angle;
-}
+float wrapAngle(float angle);
 
 #endif // WRAPPING_H
